Matchmaker restructuring for cleanliness
